requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-django==4.8.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
